// ui/src/types/bindings/entry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Entry = { uuid: string, feed_uuid: string, url: string, title: string | null, summary: string | null, content_html: string | null, published_at: string | null, updated_at: string | null, };